refactor: Menu 컨트롤러에서 도메인 제거
