+++ src/main/java/kitchenpos/menu/dto/MenuProductRequest.java
@@ -1,0 +1,36 @@
+package kitchenpos.menu.dto;
+
+import kitchenpos.menu.domain.MenuProduct;
+
+public class MenuProductRequest {
+    private Long productId;
+    private long quantity;
+
+    public MenuProductRequest() {
+    }
+
+    public MenuProductRequest(Long productId, long quantity) {
+        this.productId = productId;
+        this.quantity = quantity;
+    }
+
+    public Long getProductId() {
+        return productId;
+    }
+
+    public void setProductId(Long productId) {
+        this.productId = productId;
+    }
+
+    public long getQuantity() {
+        return quantity;
+    }
+
+    public void setQuantity(long quantity) {
+        this.quantity = quantity;
+    }
+
+    public MenuProduct toEntity() {
+        return new MenuProduct(productId, quantity);
+    }
+}
