+++ src/main/java/kitchenpos/menu/dto/MenuProductResponse.java
@@ -1,0 +1,39 @@
+package kitchenpos.menu.dto;
+
+import kitchenpos.menu.domain.MenuProduct;
+
+public class MenuProductResponse {
+    private Long seq;
+    private Long productId;
+    private long quantity;
+
+    public MenuProductResponse(MenuProduct menuProduct) {
+        seq = menuProduct.getSeq();
+        productId = menuProduct.getProductId();
+        quantity = menuProduct.getQuantity();
+    }
+
+    public Long getSeq() {
+        return seq;
+    }
+
+    public void setSeq(Long seq) {
+        this.seq = seq;
+    }
+
+    public Long getProductId() {
+        return productId;
+    }
+
+    public void setProductId(Long productId) {
+        this.productId = productId;
+    }
+
+    public long getQuantity() {
+        return quantity;
+    }
+
+    public void setQuantity(long quantity) {
+        this.quantity = quantity;
+    }
+}
