+++ src/main/java/kitchenpos/menu/dto/MenuRequest.java
@@ -1,0 +1,61 @@
+package kitchenpos.menu.dto;
+
+import kitchenpos.menu.domain.Menu;
+
+import java.util.List;
+
+public class MenuRequest {
+    private String name;
+    private long price;
+    private Long menuGroupId;
+    private List<MenuProductRequest> menuProducts;
+
+    public MenuRequest() {
+    }
+
+    public MenuRequest(String name, long price, Long menuGroupId) {
+        this.name = name;
+        this.price = price;
+        this.menuGroupId = menuGroupId;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public long getPrice() {
+        return price;
+    }
+
+    public void setPrice(long price) {
+        this.price = price;
+    }
+
+    public Long getMenuGroupId() {
+        return menuGroupId;
+    }
+
+    public void setMenuGroupId(Long menuGroupId) {
+        this.menuGroupId = menuGroupId;
+    }
+
+    public List<MenuProductRequest> getMenuProducts() {
+        return menuProducts;
+    }
+
+    public void setMenuProducts(List<MenuProductRequest> menuProducts) {
+        this.menuProducts = menuProducts;
+    }
+
+    public Menu toEntity() {
+        Menu menu = new Menu(name, price, menuGroupId);
+        menuProducts.stream()
+                .map(MenuProductRequest::toEntity)
+                .forEach(menu::addMenuProduct);
+        return menu;
+    }
+}
